Move defcon contest page into its own folder & created landing page for hack siv
