+++ app/2024defcon/SIVVoteUIScreenshot.tsx
@@ -1,7 +1,7 @@
 'use client'
 
-import sivVoteUI from '../public/images-for-decon-2024/screenshot-of-siv-voter-interface.png'
-import { scaleWithAmbientLight } from './2024defcon/scaleWithAmbientLight'
+import sivVoteUI from '../../public/images-for-decon-2024/screenshot-of-siv-voter-interface.png'
+import { scaleWithAmbientLight } from './scaleWithAmbientLight'
 import Image from 'next/image'
 
 export const SIVVoteUIScreenshot = () => (
@@ -9,7 +9,6 @@ export const SIVVoteUIScreenshot = () => (
     src={sivVoteUI}
     alt="Screenshot of SIV Vote UI"
     width={300}
-    height={300}
     style={{ opacity: scaleWithAmbientLight(0.5, 0.85) }}
   />
 )
