+++ app/2024defcon/images/dc32-logo-2color-halftonegradient.webp
