+++ app/2024defcon/images/good-luck.png
