+++ app/2024defcon/images/votingvillage.png
