+++ app/2024defcon/scaleWithAmbientLight.ts
